pyproject.toml
[tool.poetry]
name = "hopnbarley"
version = "0.1.0"
description = ""
authors = ["Katerina FuryGale@yandex.ru"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
django = ">=5.2.5,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
django-filter = ">=25.1,<26.0"
django-cors-headers = ">=4.7.0,<5.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
drf-spectacular = ">=0.28.0,<0.29.0"
pillow = ">=11.3.0,<12.0.0"
pytest-django = ">=4.11.1,<5.0.0"
python-dotenv = "^1.1.1"
unicorn = "^2.1.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.4"  # Исправлено: убрано "django"
ipython = "^9.4.0"
django-extensions = "^4.1"
django-debug-toolbar = "^6.0.0"
django-stubs = "^5.2.5"
djangorestframework-stubs = "^3.16.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
disallow_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
pretty = true
exclude = '''(?x)(
    ^venv/|
    ^\.venv/|
    ^static/|
    ^media/|
    ^\.mypy_cache/|
    ^\.pytest_cache/|
    ^\.git/|
    ^.*?/migrations/
)'''

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["orders.cart"]
warn_return_any = false

[tool.django-stubs]
django_settings_module = "config.settings"
strict_settings = true